+++ kwcoco/demo/toydata.py
@@ -10,8 +10,8 @@ Note:
 from __future__ import absolute_import, division, print_function
 
 from .toydata_video import random_single_video_dset, random_video_dset
-from .toydata_image import demodata_toy_dset
+from .toydata_image import demodata_toy_dset, demodata_toy_img
 
 
 __all__ = ['demodata_toy_dset', 'random_single_video_dset',
-           'random_video_dset']
+           'random_video_dset', 'demodata_toy_img']